Solution to ticket #91
Removed custom_iterator and refactored GenomeTree to RegionVCFIter.
Better reuse of code and don't need the weird __next__/__iter__.
Should be faster, particularly with includebeds
Found an error where TruScore precision lost from float to int casting lost optimal matches.

+++ truvari/comparisons.py
@@ -437,7 +437,7 @@ def weighted_score(sim, size, ovl):
     :rtype: float
     """
     score = (2 * sim + 1 * size + 1 * ovl) / 3.0
-    new_score = int(score / 1.333333 * 100)
+    new_score = score / 1.333333 * 100
     return new_score
